+++ application/src/components/documents/index.ts
@@ -1,5 +1,4 @@
 export { default as DocumentsTable } from './DocumentsTable';
 export { default as PersonDocuments } from './PersonDocuments';
 export { default as ProfileImageUpload } from './ProfileImageUpload';
-export { getDocumentTypeColor } from '../../utils/helper';
-export { formatFileSize } from '../../utils/formatters';
+export { getDocumentTypeColor, formatFileSize } from '../../utils';